SPOT-763.
Track if we’re running a search.  If we are during a segue, track where we’re going, and send the results to that ui view controller, if it’s the right sort.  Also, make it so those destination ui view controllers show a spinner instead of a stale list - don’t want that list jumping all over the place.